API Transbank/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
oracledb==3.1.1
python-dotenv==1.1.1
requests==2.31.0 